refactor(tests): move integrative tests into separate dir
