+++ MarkdownLinkedImagesMover.Tests/IntegrativeTests/ProgramTests.cs
@@ -1,6 +1,6 @@
 using System.IO;
 
-namespace MarkdownLinkedImagesMover.Tests;
+namespace MarkdownLinkedImagesMover.Tests.IntegrativeTests;
 
 public sealed class ProgramTests
 {
